Split FileInfoStore, PartitionStore, and RevisionUtils into their own classes
